refactor to fit scala support
