+++ src/main/scala/org/apache/sysml/api/ml/scala/LogisticRegression.scala
@@ -1,19 +1,20 @@
-package org.apache.spark.ml.systemml
+package org.apache.sysml.api.ml.scala
 
 import org.apache.spark.{ SparkContext }
-
-import org.apache.spark.sql.{ DataFrame, Row }
+import org.apache.spark.sql.DataFrame
 import org.apache.spark.sql.types.StructType
-
 import org.apache.spark.ml.{ Model, Estimator }
 import org.apache.spark.ml.classification._
-import org.apache.spark.ml.param.{ Params, Param, ParamMap }
+import org.apache.spark.ml.param.{ Params, Param, ParamMap,DoubleParam }
 import org.apache.spark.ml.param.shared._
-import org.apache.spark.mllib.linalg.Vector
-
 import org.apache.sysml.api.{ MLContext, MLOutput }
 import org.apache.sysml.runtime.matrix.MatrixCharacteristics
 import org.apache.sysml.runtime.instructions.spark.utils.{ RDDConverterUtilsExt => RDDConverterUtils }
+import org.apache.spark.SparkConf
+import org.apache.spark.mllib.linalg.Vectors
+import org.apache.spark.mllib.regression.LabeledPoint
+import org.apache.sysml.runtime.instructions.spark.utils.{RDDConverterUtilsExt => RDDConverterUtils}
+
 
 trait HasIcpt extends Params {
   final val icpt: Param[Int] = new Param[Int](this, "icpt", "Intercept presence, shifting and rescaling X columns")
@@ -30,14 +31,22 @@ trait HasMaxInnerIter extends Params {
   setDefault(maxInnerIter, 0)
   final def getMaxInnerIter: Int = $(maxInnerIter)
 }
+trait HasTol extends Params {
+  final val tol: DoubleParam = new DoubleParam(this, "tol", "the convergence tolerance for iterative algorithms")
+  setDefault(tol,0.000001)
+  final def getTol: Double = $(tol)
+}
+trait HasRegParam extends Params {
+  final val regParam: DoubleParam = new DoubleParam(this, "tol", "the convergence tolerance for iterative algorithms")
+  setDefault(regParam,0.000001)
+  final def getRegParam: Double = $(regParam)
+}
 object LogisticRegression{
   final val scriptPath = "MultiLogReg.dml"
 }
 class LogisticRegression(override val uid: String,val sc:SparkContext) extends Estimator[LogisticRegressionModel] with HasIcpt
     with HasRegParam with HasTol with HasMaxOuterIter with HasMaxInnerIter {
   
-  setDefault(regParam, 0.0001)
-  setDefault(tol, 0.0001)
   def setIcpt(value: Int) = set(icpt, value)
   def setMaxOuterIter(value: Int) = set(maxOuterIter, value)
   def setMaxInnerIter(value: Int) = set(maxInnerIter, value)
@@ -107,14 +116,15 @@ class LogisticRegressionModel(
 object LogisticRegressionTest {
   import org.apache.spark.{ SparkConf, SparkContext }
   import org.apache.spark.sql.types._
-  import org.apache.spark.mllib.linalg.Vectors
-  import org.apache.spark.mllib.regression.LabeledPoint
+import org.apache.spark.mllib.linalg.Vectors
+import org.apache.spark.mllib.regression.LabeledPoint
   
   def main(args: Array[String]) = {
     val sparkConf: SparkConf = new SparkConf();
     val sc: SparkContext = new SparkContext("local", "TestLocal", sparkConf);
     val sqlContext = new org.apache.spark.sql.SQLContext(sc);
-    import sqlContext.implicits._
+
+import sqlContext.implicits._
     ScriptsUtils.setSystemmlHome("C:\\aWorkFolder\\git\\systemml\\incubator-systemml\\scripts")
     val training = sc.parallelize(Seq(
       LabeledPoint(1.0, Vectors.dense(1.0, 0.0, 3.0)),
